card design tweaks, code organization
